Bash environment configuration
At first, bash seems to be a nicely designed piece of software, but with a
closer look one can clearly see a lot of duck tape and WD40. For instance
it is NOT possible to correctly organize .profile and .bashrc files. Every
time something is missing from login shell or from the interactive shell.
The funny thing is, that the login shell can be interactive...

Even in Gentoo this configuration is fucked up beyond possible repair. The
problem is, that /etc/profile is doing stuff which bashrc should be doing
(e.g. loading bash completions). At the end of the day, all bash instances
are forced to load /etc/profile.

So, the long story short. In order to make everything reasonably "OK", one
has to create configuration in which login shell and interactive shell are
loading the same files - there is no separation between those shells.